+++ NEWS.md
@@ -4,6 +4,8 @@
 - The `--precompile` elaboration option which was deprecated in the last
   release has been removed.  The related `--with-system-cc` configure
   option has also been removed.
+- The `--stderr=none` option now prevents any diagnostic messages being
+  printed to `stderr`, including those at `failure` severity (#1357).
 - Several other minor bugs were resolved (#1237, #1351).
 
 ## Version 1.18.2 - 2025-11-19